backend/requirements.txt
Django>=5.0.0,<6.0.0
djangorestframework>=3.15.0,<4.0.0
django-cors-headers>=4.3.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
djangorestframework-simplejwt>=5.3.0,<6.0.0
PyJWT>=2.8.0,<3.0.0

# Database
psycopg2-binary>=2.9.9,<3.0.0

# Cache and Task Queue
redis>=5.0.0,<6.0.0
django-redis>=5.4.0,<6.0.0

# API Documentation
drf-spectacular>=0.27.0,<1.0.0

# Google Gemini AI
google-generativeai>=0.3.0,<1.0.0

# Environment and Configuration
django-environ>=0.11.0,<1.0.0

# Utilities
requests>=2.31.0,<3.0.0 